Introduce binary resources renderer callback
Although implemented and tested, there is no code-path that uses this callback from the plugin.
